implement the CQGLProgram interface.

+++ CQEngine/Source/Mathematics/CQTuple.inl
@@ -4,7 +4,7 @@ Tuple<DIMENSION,TYPE>::Tuple()
 
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 Tuple<DIMENSION, TYPE>::Tuple(const Tuple& _tuple)
 {
 	for (int i = 0; i < DIMENSION; ++i)
@@ -13,41 +13,41 @@ Tuple<DIMENSION, TYPE>::Tuple(const Tuple& _tuple)
 	}
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 Tuple<DIMENSION, TYPE>::~Tuple()
 {
 
 }
 
 // *ptr
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 inline Tuple<DIMENSION, TYPE>::operator const TYPE* () const
 {
 	return mTuple;
 }
 
 // *ptr
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 inline Tuple<DIMENSION, TYPE>::operator TYPE* ()
 {
 	return mTuple;
 }
 
 // tuple[i]
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 inline const TYPE& Tuple<DIMENSION, TYPE>::operator[] (int _i) const
 {
 	return mTuple[_i];
 }
 
 // tuple[i]
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 inline TYPE& Tuple<DIMENSION, TYPE>::operator[] (int _i)
 {
 	return mTuple[_i];
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 Tuple<DIMENSION, TYPE>& Tuple<DIMENSION, TYPE>::operator= (const Tuple& _tuple)
 {
 	for (int i = 0; i < DIMENSION; ++i)
@@ -57,7 +57,7 @@ Tuple<DIMENSION, TYPE>& Tuple<DIMENSION, TYPE>::operator= (const Tuple& _tuple)
 	return *this;
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 bool Tuple<DIMENSION, TYPE>::operator== (const Tuple& _tuple) const
 {
 	for (int i = 0; i < DIMENSION; ++i)
@@ -70,13 +70,13 @@ bool Tuple<DIMENSION, TYPE>::operator== (const Tuple& _tuple) const
 	return true;
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 bool Tuple<DIMENSION, TYPE>::operator!= (const Tuple& _tuple) const
 {
 	return !(*this == _tuple);
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 bool Tuple<DIMENSION, TYPE>::operator< (const Tuple& _tuple) const
 {
 	for (int i = 0; i < DIMENSION; ++i)
@@ -94,7 +94,7 @@ bool Tuple<DIMENSION, TYPE>::operator< (const Tuple& _tuple) const
 	return false;
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 bool Tuple<DIMENSION, TYPE>::operator<= (const Tuple& _tuple) const
 {
 	for (int i = 0; i < DIMENSION; ++i)
@@ -112,7 +112,7 @@ bool Tuple<DIMENSION, TYPE>::operator<= (const Tuple& _tuple) const
 	return true;
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 bool Tuple<DIMENSION, TYPE>::operator> (const Tuple& _tuple) const
 {
 	for (int i = 0; i < DIMENSION; ++i)
@@ -130,7 +130,7 @@ bool Tuple<DIMENSION, TYPE>::operator> (const Tuple& _tuple) const
 	return false;
 }
 
-template <int DIMENSION, typename TYPE>
+template <size_t DIMENSION, typename TYPE>
 bool Tuple<DIMENSION, TYPE>::operator>= (const Tuple& _tuple) const
 {
 	for (int i = 0; i < DIMENSION; ++i)
